backend/requirements.txt
Flask==2.2.2
Flask-CORS==3.0.10
mysql-connector-python==9.1
python-dotenv==0.19
requests==2.32
Werkzeug==2.2.2